Mini Project
In this mini-project bricks breaker game is implemented by using OpenGL APIs and computer graphics concepts. When we start the bricks breaker game initially the gun is in a constant position. To make the movement of the gun, they used different keyboard functions and display to make the program user interactive.

The rectangle shape of bricks created outside the window the bricks move down. If we use only keyboard function then the only bullet will fire. If gunfire the bullet exact center of the bricks then collision of the bricks takes place. After the collision of bricks, the score will be increased.

When red and yellow bricks are collided by bullet same bricks will come by a distance of +50 on the x-axis. If blue and green bricks are collided by bullet same bricks will come by a distance of -50 on the x-axis. When the bricks reach the window game over is displayed with the current score.

By using the keyboard interface, we perform the required action. The functionality of various keyboard keys is:

     f – press key f for the Instructions page

     n – press n to start the Game

     f/F – Bullet from gun fires by either pressing key for F once the game starts.

     a/A – Gun move left side by pressing either key a or A

    d/D – Gun move right side by pressing either key d or D

    q – press key q to Exit the game